my-app/backend/requirements.txt
fastapi
uvicorn[standard]
supabase
openai
python-dotenv
pydantic
